xadmin_requirements.txt
six
future
httplib2
django-reversion
django-formtools
django-crispy-forms
django-import-export